requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==1.0.0
Flask-Bcrypt==0.7.1
Flask-Migrate==3.1.0
python-dotenv==0.19.2
Werkzeug==2.0.3